# projects/agent-kanban/docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Agent Kanban API
  description: |
    多项目 AI Agent 统一看板的 REST API。

    核心资源链：Workspace → Project → Task → Assignment → Agent

    **契约说明**：
    - Task 的创建/更新/关闭以 GitHub Issues 为准，本 API 提供只读聚合查询。
    - Assignment 的创建和管理通过本 API。
    - 此文件是 **canonical** API 契约，Go 和 Node.js 两个后端都基于此实现。
    - 所有时间字段为 ISO 8601 格式 (e.g. "2026-08-03T07:22:00Z")。
  version: "1.1.0"
  contact:
    name: Agent Kanban Team
servers:
  - url: http://localhost:8080/api/v1
    description: Go 后端 (本地开发)
  - url: http://localhost:4001/api/v1
    description: Node.js 后端 (本地开发)
  - url: http://localhost:4002/api/v1
    description: Next.js 嵌入式后端 (本地开发)
security:
  - ApiKeyAuth: []
  - {}
...
paths:
...
  /workspaces/{workspaceId}/agents:
    get:
      tags: [Agents]
      summary: 获取工作空间的所有 Agent
      operationId: listAgents
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: Agent 列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Agent"
    post:
      tags: [Agents]
      summary: 注册新 Agent
      operationId: createAgent
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgentInput"
            examples:
              frontend:
                summary: 前端 Agent
                value:
                  name: "前端开发专家"
                  role: "frontend"
                  description: "负责 React/Vue 前端开发"
                  modelName: "kimi-long-v1"
                  capabilityTags: ["frontend", "ui", "component"]
                  skills: ["react-component", "tailwind-css"]
                  maxConcurrency: 3
              backend:
                summary: 后端 Agent
                value:
                  name: "后端开发专家"
                  role: "backend"
                  description: "负责 API 和数据库设计"
                  modelName: "glm-4-flash"
                  capabilityTags: ["backend", "api", "database"]
                  maxConcurrency: 3
      responses:
        "201":
          description: 注册成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "400":
          $ref: "#/components/responses/ValidationError"
  /agents/{agentId}:
    get:
      tags: [Agents]
      summary: 获取 Agent 详情（含统计数据）
      operationId: getAgent
      parameters:
        - $ref: "#/components/parameters/AgentId"
      responses:
        "200":
          description: Agent 详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentDetail"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Agents]
      summary: 更新 Agent 配置
      operationId: updateAgent
      parameters:
        - $ref: "#/components/parameters/AgentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAgentInput"
            examples:
              pause:
                summary: 暂停 Agent
                value:
                  status: "paused"
              update_model:
                summary: 更换模型
                value:
                  modelName: "glm-4-plus"
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Agents]
      summary: 移除 Agent
      operationId: deleteAgent
      parameters:
        - $ref: "#/components/parameters/AgentId"
      responses:
        "204":
          description: 移除成功
        "404":
          $ref: "#/components/responses/NotFound"

  # ═══════════════════════════════════════════════
  # Assignments
  # ═══════════════════════════════════════════════
  /assignments:
    post:
      tags: [Assignments]
      summary: 手动创建任务分配
      operationId: createAssignment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssignmentInput"
            examples:
              default:
                value:
                  taskId: "task-001"
                  agentId: "agent-001"
                  instruction: "请实现登录页面，使用 React + Tailwind"
      responses:
        "201":
          description: 分配创建成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
              examples:
                default:
                  value:
                    id: "asg-001"
                    taskId: "task-001"
                    agentId: "agent-001"
                    status: "pending"
                    modelUsed: null
                    resultSummary: null
                    instruction: "请实现登录页面，使用 React + Tailwind"
                    githubCommentId: null
                    retryCount: 0
                    assignedAt: "2026-08-03T07:22:00Z"
                    startedAt: null
                    completedAt: null
        "409":
          $ref: "#/components/responses/Conflict"
  /assignments/auto:
    post:
      tags: [Assignments]
      summary: 自动分配（Label Router 匹配）
      description: |
        根据 task.labels 自动匹配 capability_tags 有交集的 Agent。
        匹配标签越多优先级越高。无匹配时返回 422。
      operationId: autoAssign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [taskId]
              properties:
                taskId:
                  type: string
            examples:
              default:
                value:
                  taskId: "task-001"
      responses:
        "201":
          description: 自动匹配成功，分配已创建
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
        "422":
          description: 没有匹配的 Agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                no_match:
                  value:
                    code: "NO_MATCHING_AGENT"
                    message: "No agent matches labels: [unknown-type]"
  /tasks/{taskId}/assignments:
    get:
      tags: [Assignments]
      summary: 获取任务的所有分配记录
      operationId: listTaskAssignments
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "200":
          description: 分配记录列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assignment"
  /agents/{agentId}/assignments:
    get:
      tags: [Assignments]
      summary: 获取 Agent 的所有分配记录
      operationId: listAgentAssignments
      parameters:
        - $ref: "#/components/parameters/AgentId"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/AssignmentStatus"
      responses:
        "200":
          description: 分配记录列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assignment"

  # ═══════════════════════════════════════════════
  # Webhooks
  # ═══════════════════════════════════════════════
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: 可选。Zero-Auth 模式下不需要。生产环境启用时在 header 传 X-API-Key。
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema:
        type: string
...
    TaskId:
      name: taskId
      in: path
      required: true
      schema:
        type: string
    AgentId:
      name: agentId
      in: path
      required: true
      schema:
        type: string
  responses:
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            default:
              value:
                code: "NOT_FOUND"
                message: "Workspace not found"
    ValidationError:
      description: 请求参数验证失败
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            default:
              value:
                code: "VALIDATION_ERROR"
                message: "name is required"
                details:
                  field: "name"
    Conflict:
      description: 资源冲突
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            duplicate_project:
              value:
                code: "CONFLICT"
                message: "Project with repo my-org/frontend-app already exists"
            duplicate_assignment:
              value:
                code: "CONFLICT"
                message: "Task already has a pending or running assignment"
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum: [NOT_FOUND, VALIDATION_ERROR, CONFLICT, UNAUTHORIZED, NO_MATCHING_AGENT, INTERNAL_ERROR]
        message:
          type: string
        details:
          type: object
...
    AssignmentStatus:
      type: string
      enum: [pending, running, done, failed]
...
    Agent:
      type: object
      required: [id, workspaceId, name, role, modelName, capabilityTags, status, maxConcurrency]
      properties:
        id:
          type: string
        workspaceId:
          type: string
        name:
          type: string
        role:
          type: string
        description:
          type: string
        modelName:
          type: string
        capabilityTags:
          type: array
          items:
            type: string
        skills:
          type: array
          items:
            type: string
        githubBotAccount:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, paused, offline]
        maxConcurrency:
          type: integer
          default: 3
        createdAt:
          type: string
          format: date-time
    CreateAgentInput:
      type: object
      required: [name, role, modelName, capabilityTags]
      properties:
        name:
          type: string
        role:
          type: string
        description:
          type: string
        modelName:
          type: string
        capabilityTags:
          type: array
          items:
            type: string
        skills:
          type: array
          items:
            type: string
        githubBotAccount:
          type: string
        maxConcurrency:
          type: integer
          default: 3
    UpdateAgentInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        modelName:
          type: string
        capabilityTags:
          type: array
          items:
            type: string
        skills:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, paused, offline]
        maxConcurrency:
          type: integer
...
    AgentDetail:
      allOf:
        - $ref: "#/components/schemas/Agent"
        - type: object
          properties:
            stats:
              type: object
              properties:
                totalAssignments:
                  type: integer
                completedAssignments:
                  type: integer
                failedAssignments:
                  type: integer
                runningCount:
                  type: integer
                avgDurationMs:
                  type: integer
                  nullable: true
    Assignment:
      type: object
      required: [id, taskId, agentId, status, retryCount, assignedAt]
      properties:
        id:
          type: string
        taskId:
          type: string
        agentId:
          type: string
        status:
          $ref: "#/components/schemas/AssignmentStatus"
        modelUsed:
          type: string
          nullable: true
        resultSummary:
          type: string
          nullable: true
        instruction:
          type: string
          nullable: true
        githubCommentId:
          type: integer
          nullable: true
        retryCount:
          type: integer
        assignedAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        agent:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            role:
              type: string
            modelName:
              type: string
    CreateAssignmentInput:
      type: object
      required: [taskId, agentId]
      properties:
        taskId:
          type: string
        agentId:
          type: string
        instruction:
          type: string